+++ changelog.md
@@ -1,6 +1,11 @@
 # Changelog
 
-## [1.1.2] - 27.04.2025
+## [1.2.0] - 27.04.2025
+
+### Added
+
+- Added output of generated settings link
+- Added option to show settings link in plugin list
 
 ### Changed